gui for mouse atlas with magicclass and magicgui
